ablog/requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-ckeditor==6.7.0
django-filter==2.3.0
django-import-export==3.3.3
django-js-asset==2.1.0
django-simple-history==3.4.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.1.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
